backend/requirements.txt
# Python dependencies
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.1.0